+++ critical_failure/lib/src/pages/player_subpages/player_page_clases.dart
@@ -1,5 +1,6 @@
+import 'package:critical_failure/src/pages/player_page.dart';
+import 'package:critical_failure/src/pages/player_subpages/drawer_builder.dart';
 import 'package:flutter/material.dart';
-
 class PlayerClases extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -7,6 +8,7 @@ class PlayerClases extends StatelessWidget {
       appBar: AppBar(
         title: Text('Clases'),
       ),
+      drawer: DrawerBuilder(),
     );
   }
 }
