+++ critical_failure/lib/src/pages/player_subpages/player_page_razas.dart
@@ -1,3 +1,4 @@
+import 'package:critical_failure/src/pages/player_subpages/drawer_builder.dart';
 import 'package:flutter/material.dart';
 
 class PlayerRazas extends StatelessWidget {
@@ -7,6 +8,7 @@ class PlayerRazas extends StatelessWidget {
       appBar: AppBar(
         title: Text('Razas'),
       ),
+      drawer: DrawerBuilder(),
     );
   }
