+++ critical_failure/lib/src/pages/player_subpages/player_page_trasfondo.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'drawer_builder.dart';
+
 class PlayerTrasfondo extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -7,6 +9,7 @@ class PlayerTrasfondo extends StatelessWidget {
       appBar: AppBar(
         title: Text('Trasfondo'),
       ),
+      drawer: DrawerBuilder(),
     );
   }
